fix names of functions/variables/constants to be a good Go citizen
